+++ terraform/slurm_cluster/examples/slurm_cluster/hybrid/full/variables.tf
@@ -142,6 +142,39 @@ variable "cloud_parameters" {
   }
 }
 
+variable "disable_default_mounts" {
+  description = <<-EOD
+    Disable default global network storage from the controller
+    * /usr/local/etc/slurm
+    * /etc/munge
+    * /home
+    * /apps
+    If these are disabled, the slurm etc and munge dirs must be added manually,
+    or some other mechanism must be used to synchronize the slurm conf files
+    and the munge key across the cluster.
+    EOD
+  default     = false
+}
+
+variable "network_storage" {
+  description = <<EOD
+Storage to mounted on all instances.
+* server_ip     : Address of the storage server.
+* remote_mount  : The location in the remote instance filesystem to mount from.
+* local_mount   : The location on the instance filesystem to mount to.
+* fs_type       : Filesystem type (e.g. "nfs").
+* mount_options : Options to mount with.
+EOD
+  type = list(object({
+    server_ip     = string
+    remote_mount  = string
+    local_mount   = string
+    fs_type       = string
+    mount_options = string
+  }))
+  default = []
+}
+
 variable "slurmdbd_conf_tpl" {
   description = "Slurm slurmdbd.conf template file path."
   type        = string
@@ -217,12 +250,14 @@ Variables map to:
 EOD
   type = object({
     google_app_cred_path = string
+    slurm_control_host   = string
     slurm_bin_dir        = string
     slurm_log_dir        = string
     output_dir           = string
   })
   default = {
     google_app_cred_path = null
+    slurm_control_host   = null
     slurm_bin_dir        = "/usr/local/bin"
     slurm_log_dir        = "/var/log/slurm"
     output_dir           = "."