+++ terraform/slurm_cluster/main.tf
@@ -126,6 +126,7 @@ module "slurm_controller_instance" {
   enable_reconfigure           = var.enable_reconfigure
   epilog_scripts               = var.epilog_scripts
   login_network_storage        = var.login_network_storage
+  disable_default_mounts       = var.disable_default_mounts
   network_storage              = var.network_storage
   partitions                   = values(module.slurm_partition)[*]
   prolog_scripts               = var.prolog_scripts
@@ -146,6 +147,7 @@ module "slurm_controller_hybrid" {
   slurm_cluster_name = var.slurm_cluster_name
 
   google_app_cred_path         = var.controller_hybrid_config.google_app_cred_path
+  slurm_control_host           = var.controller_hybrid_config.slurm_control_host
   slurm_bin_dir                = var.controller_hybrid_config.slurm_bin_dir
   slurm_log_dir                = var.controller_hybrid_config.slurm_log_dir
   output_dir                   = var.controller_hybrid_config.output_dir