+++ terraform/slurm_cluster/modules/slurm_controller_hybrid/variables.tf
@@ -23,16 +23,6 @@ variable "project_id" {
   description = "Project ID to create resources in."
 }
 
-############
-# METADATA #
-############
-
-variable "metadata" {
-  type        = map(string)
-  description = "Metadata key/value pairs."
-  default     = null
-}
-
 #########
 # SLURM #
 #########
@@ -140,6 +130,20 @@ EOD
   default = []
 }
 
+variable "disable_default_mounts" {
+  description = <<-EOD
+    Disable default global network storage from the controller
+    * /usr/local/etc/slurm
+    * /etc/munge
+    * /home
+    * /apps
+    If these are disabled, the slurm etc and munge dirs must be added manually,
+    or some other mechanism must be used to synchronize the slurm conf files
+    and the munge key across the cluster.
+    EOD
+  default     = false
+}
+
 variable "network_storage" {
   description = <<EOD
 Storage to mounted on all instances.
@@ -239,6 +243,17 @@ variable "slurm_log_dir" {
   type        = string
   description = "Directory where Slurm logs to."
   default     = "/var/log/slurm"
+}
+
+variable "slurm_control_host" {
+  type        = string
+  description = <<EOD
+The short, or long, hostname of the machine where Slurm control daemon is
+executed (i.e. the name returned by the command "hostname -s").
+
+See https://slurm.schedmd.com/slurm.conf.html#OPT_SlurmctldHost
+EOD
+  default     = null
 }
 
 variable "cloud_parameters" {